==> notebooks/00_data_split.ipynb <==
import pandas as pd

df = pd.read_csv(r'C:\Users\Paweł\Desktop\MLRegression\MLR\data\raw\untouched_raw_original.csv')

pd.set_option('display.max_columns', None)  # Show all columns
pd.set_option('display.max_rows', None)     # Show all rows

df.shape

df.head(2)

df['date'] = pd.to_datetime(df['date'])
print(df['date'].dt.year.value_counts())

##### Avoid Leakage

- Your train set (2012–2019) is “clean” historical training.

- Your eval set (2020–2021) is “unseen” during training, but still used for validation and hyperparameter tuning.

- Your holdout set (2022–2023) is completely untouched until the very end.

# Sort by date (important to avoid shuffling issues)
df = df.sort_values("date")

# Define cutoff dates
cutoff_date_eval = "2020-01-01"   # validation starts
cutoff_date_holdout = "2022-01-01"  # holdout starts

# Train: before 2020
train_df = df[df["date"] < cutoff_date_eval]

# Validation/Eval: 2020–2021
eval_df = df[(df["date"] >= cutoff_date_eval) & (df["date"] < cutoff_date_holdout)]

# Holdout: 2022–2023
holdout_df = df[df["date"] >= cutoff_date_holdout]

print("Train shape:", train_df.shape)
print("Eval shape:", eval_df.shape)
print("Holdout shape:", holdout_df.shape)

train_df.head(2)

eval_df.head(2)

holdout_df.head(2)

# Save splits
train_df.to_csv(r'C:\Users\Paweł\Desktop\MLRegression\MLR\data\raw\train.csv', index=False)
eval_df.to_csv(r'C:\Users\Paweł\Desktop\MLRegression\MLR\data\raw\eval.csv', index=False)
holdout_df.to_csv(r'C:\Users\Paweł\Desktop\MLRegression\MLR\data\raw\holdout.csv', index=False)
